Move non-FITS functions to a utils module
